Completing the application by adding the company and Review functionalities

+++ src/main/java/com/madz/firstJobApp/job/JobService.java
@@ -7,7 +7,7 @@ public interface JobService {
     List<Job>  findAll();
     void createJob(Job job);
 
-    Job jobById(int id);
+    Job jobById(long id);
 
     boolean deleteById(long id);
